pyproject.toml
[project]
name = "truthapi"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.129.0",
    "pydantic>=2.12.5",
    "pyyaml>=6.0.3",
    "uvicorn>=0.40.0",
]
